btrfs-progs: convert: widen int types in convert context
The u32 types in the convert context might not be enough for some very
large filesytems (20TB). Use 64bit types to be safe.

+++ convert/common.h
@@ -30,10 +30,10 @@ struct btrfs_mkfs_config;
 
 struct btrfs_convert_context {
 	u32 blocksize;
-	u32 first_data_block;
-	u32 block_count;
-	u32 inodes_count;
-	u32 free_inodes_count;
+	u64 first_data_block;
+	u64 block_count;
+	u64 inodes_count;
+	u64 free_inodes_count;
 	u64 total_bytes;
 	char *volume_name;
 	const struct btrfs_convert_operations *convert_ops;
